Zones kinda fixed, Need to add cave and basement to traveable. moved info into Data
